backend/requirements.txt
# ClaimSense.ai - Backend Dependencies

# Web framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.12

# Data validation
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0

# LLM - Google Gemini (official SDK, replaces deprecated google-generativeai)
google-genai>=1.0.0

# PDF text extraction
PyPDF2>=3.0.0
pdfplumber>=0.11.0

# OCR
pytesseract>=0.3.13
Pillow>=10.0.0

# PDF to image (for OCR tier)
pdf2image>=1.17.0

# Encryption (PII field protection)
cryptography>=43.0.0

# Date parsing (for M2 waiting period validation)
python-dateutil>=2.9.0

# Database
psycopg2-binary>=2.9.0,<3.0.0

# Environment
python-dotenv>=1.0.0
